// src/config/interfaces/Envs.interfaces.ts
export interface EnvsInterface {
    port: number;
    server_status: string;
    server_key: string;
    secret_key: string;
    salt: string;
    db_dialect: string;
    db_port: number;
    db_name: string;
    db_user: string;
    db_pass: string;
    db_host: string;
    basic_path: string;
  }
  